remove BUFFER_SIZE from file_descriptor.hh
Previously, FileDescriptor::BUFFER_SIZE was declared in file_descriptor.hh
but not defined in file_descriptor.cc. This could cause errors linking
the example programs against libsourdough.a. (This error is masked when
libsourdough is built with -O2 because FileDescriptor::BUFFER_SIZE gets
const-folded away. But building with, say, -Og would cause breakage.)

This commit makes BUFFER_SIZE a constexpr in the read() definition, and
changes the default value of limit to numeric_limits<size_t>::max().
Since the read size is min(limit, BUFFER_SIZE), this default is safe.

Tested with g++ 8.2.1 and clang++ 7.0.0.